# WIP: Implement DRA admin-controlled attributes (KEP-5027)