--- a/pkg/apis/resource/register.go
+++ b/pkg/apis/resource/register.go
@@ -60,6 +60,8 @@ func addKnownTypes(scheme *runtime.Scheme) error {
 		&ResourceClaimTemplateList{},
 		&ResourceSlice{},
 		&ResourceSliceList{},
+		&ResourceSlicePatch{},
+		&ResourceSlicePatchList{},
 	)
 
 	return nil