--- a/staging/src/k8s.io/api/resource/v1alpha3/register.go
+++ b/staging/src/k8s.io/api/resource/v1alpha3/register.go
@@ -52,6 +52,8 @@ func addKnownTypes(scheme *runtime.Scheme) error {
 		&ResourceClaimTemplateList{},
 		&ResourceSlice{},
 		&ResourceSliceList{},
+		&ResourceSlicePatch{},
+		&ResourceSlicePatchList{},
 	)
 
 	// Add the watch version that applies