… api/discovery/aggregated_v2.json
… api/discovery/apis__resource.k8s.io__v1alpha3.json
… api/openapi-spec/swagger.json
… api/openapi-spec/v3/apis__resource.k8s.io__v1alpha3_openapi.json
… api/openapi-spec/v3/apis__resource.k8s.io__v1beta1_openapi.json
… pkg/apis/resource/v1alpha3/zz_generated.conversion.go
… pkg/apis/resource/validation/validation_resourceslicepatch_test.go
… pkg/apis/resource/zz_generated.deepcopy.go
… pkg/generated/openapi/zz_generated.openapi.go
… pkg/registry/resource/resourceslicepatch/storage/storage_test.go
… pkg/registry/resource/resourceslicepatch/strategy_test.go
… pkg/scheduler/eventhandlers_test.go
… pkg/scheduler/framework/autoscaler_contract/lister_contract_test.go
… pkg/scheduler/framework/plugins/dynamicresources/dra_manager.go
… pkg/scheduler/framework/plugins/dynamicresources/dynamicresources_test.go
… pkg/scheduler/scheduler.go
… plugin/pkg/auth/authorizer/rbac/bootstrappolicy/policy.go
… staging/src/k8s.io/api/resource/v1alpha3/generated.pb.go
… staging/src/k8s.io/api/resource/v1alpha3/generated.proto
… staging/src/k8s.io/api/resource/v1alpha3/types.go
… staging/src/k8s.io/api/resource/v1alpha3/types_swagger_doc_generated.go
… staging/src/k8s.io/api/resource/v1alpha3/zz_generated.deepcopy.go
… staging/src/k8s.io/api/resource/v1alpha3/zz_generated.prerelease-lifecycle.go
… staging/src/k8s.io/api/resource/v1beta1/generated.proto
… staging/src/k8s.io/api/resource/v1beta1/types.go
… staging/src/k8s.io/api/resource/v1beta1/types_swagger_doc_generated.go
… staging/src/k8s.io/api/testdata/HEAD/resource.k8s.io.v1alpha3.ResourceSlicePatch.json
… staging/src/k8s.io/api/testdata/HEAD/resource.k8s.io.v1alpha3.ResourceSlicePatch.pb
… staging/src/k8s.io/api/testdata/HEAD/resource.k8s.io.v1alpha3.ResourceSlicePatch.yaml
… staging/src/k8s.io/client-go/applyconfigurations/internal/internal.go