@nojnhuh 2025-02-12
I think this PR technically ticks all the boxes in the KEP, but it could use a fair bit of polish around performance optimization, error handling, and documentation. I think I'll need help with some of that, so I'll make a pass at adding some of my own commentary inline.