@nojnhuh 2025-02-24
I've done a couple passes to pick off the low hanging fruit to improve performance, so I think overall this PR is ready for some early feedback. There's some polish I plan to add (comments, logs, etc.) that I'll work on in the meantime.

@pohly (or anyone else) I can also rebase this and clean up the history here, so let me know when that would be helpful and not too disruptive.

<summary>
Benchmark progress over the last week:

<details>

```
% benchstat jon/{before,after}.txt
goos: linux
goarch: amd64
pkg: k8s.io/dynamic-resource-allocation/resourceslice/tracker
cpu: Intel(R) Xeon(R) Platinum 8370C CPU @ 2.80GHz
                                                                │ jon/before.txt  │            jon/after.txt            │
                                                                │     sec/op      │   sec/op     vs base                │
EventHandlers/one-patch-to-many-slices-add-patch-16                   97.83m ± 1%   22.03m ± 1%  -77.48% (p=0.000 n=10)
EventHandlers/one-patch-to-many-slices-add-slice-16                  189.70µ ± 0%   33.76µ ± 1%  -82.20% (p=0.000 n=10)
EventHandlers/one-patched-device-among-many-slices-add-patch-16   32364.229µ ± 1%   5.072µ ± 0%  -99.98% (p=0.000 n=10)
EventHandlers/one-patched-device-among-many-slices-add-slice-16      66.996µ ± 1%   3.679µ ± 2%  -94.51% (p=0.000 n=10)
EventHandlers/one-patch-for-each-of-many-slices-add-patch-16      103781.87µ ± 0%   60.65µ ± 2%  -99.94% (p=0.000 n=10)
EventHandlers/resource-slice-add-no-patches-16                      61222.5n ± 0%   485.6n ± 1%  -99.21% (p=0.000 n=10)
geomean                                                               2.519m        27.24µ       -98.92%

                                                                │  jon/before.txt  │            jon/after.txt             │
                                                                │       B/op       │     B/op      vs base                │
EventHandlers/one-patch-to-many-slices-add-patch-16                   69.00Mi ± 0%   36.76Mi ± 0%  -46.72% (p=0.000 n=10)
EventHandlers/one-patch-to-many-slices-add-slice-16                  139.17Ki ± 0%   75.40Ki ± 0%  -45.82% (p=0.000 n=10)
EventHandlers/one-patched-device-among-many-slices-add-patch-16   23643.761Ki ± 0%   4.578Ki ± 0%  -99.98% (p=0.000 n=10)
EventHandlers/one-patched-device-among-many-slices-add-slice-16      49.310Ki ± 0%   4.523Ki ± 0%  -90.83% (p=0.000 n=10)
EventHandlers/one-patch-for-each-of-many-slices-add-patch-16       76526.04Ki ± 0%   87.43Ki ± 0%  -99.89% (p=0.000 n=10)
EventHandlers/resource-slice-add-no-patches-16                        46750.5 ± 0%     256.0 ± 0%  -99.45% (p=0.000 n=10)
geomean                                                               1.806Mi        32.97Ki       -98.22%

                                                                │ jon/before.txt │            jon/after.txt            │
                                                                │   allocs/op    │  allocs/op   vs base                │
EventHandlers/one-patch-to-many-slices-add-patch-16                  658.2k ± 0%   164.0k ± 0%  -75.08% (p=0.000 n=10)
EventHandlers/one-patch-to-many-slices-add-slice-16                  1285.0 ± 0%    331.0 ± 0%  -74.24% (p=0.000 n=10)
EventHandlers/one-patched-device-among-many-slices-add-patch-16   121659.00 ± 0%    82.00 ± 0%  -99.93% (p=0.000 n=10)
EventHandlers/one-patched-device-among-many-slices-add-slice-16      264.00 ± 0%    79.00 ± 0%  -70.08% (p=0.000 n=10)
EventHandlers/one-patch-for-each-of-many-slices-add-patch-16       656457.0 ± 0%    334.0 ± 0%  -99.95% (p=0.000 n=10)
EventHandlers/resource-slice-add-no-patches-16                      237.000 ± 0%    7.000 ± 0%  -97.05% (p=0.000 n=10)
geomean                                                              12.72k         306.1       -97.59%
```

</details>

</summary>

Removing WIP:
/retitle Implement DRA admin-controlled attributes (KEP-5027)