@nojnhuh 2025-02-25
> /label api-review

The new API here matches the [merged KEP](https://github.com/kubernetes/enhancements/blob/758ea034908515a934af09d03a927b24186af04c/keps/sig-scheduling/5027-dra-admin-controlled-device-attributes/README.md#api) with the exception of the type for ResourceSlicePatch.Spec.Priority, which I made `*int32` here instead of `*int` per the [API conventions](https://github.com/kubernetes/community/blob/e7853d897c0bf2db588e994cf07705d1148419fd/contributors/devel/sig-architecture/api-conventions.md#primitive-types).