@tico88612 2025-02-26
Hi folks,
We'd like to remind you of the code freeze time.
The code freeze is starting 02:00 UTC Friday 21st March 2025 (about 5 weeks from now). Please make sure the PR has both lgtm and approved labels before the code freeze. Thanks!