@liggitt 2025-03-13
> I don't recall reading this KEP for API -- who did that?

I'm not sure... maybe https://github.com/kubernetes/enhancements/pull/5034#discussion_r1936162686 was the only interaction?

I'm going to route this to @thockin in the API review board but as backlog (and add @soltysh as shadow since he's stepping into doing API reviews and left some API-directed comments on the KEP PR - https://github.com/kubernetes/enhancements/pull/5034/files#r1952796598)... I'll leave this to @pohly @johnbelamaric @thockin to untangle how to proceed with the API review bit