@johnbelamaric 2025-03-13
@pohly we are no longer planning on this in 1.33, correct? Instead we are updating the device taints KEP to no longer depend on it?