@pohly 2025-03-13
We already agreed to not pursue this for 1.33 - let's simply close it to avoid further confusion.

/close