@pohly 2025-03-13
Yes, https://github.com/kubernetes/kubernetes/pull/130447 is alive, kicking, and most importantly, no longer depends on this one here.